feat: 몬스터 전반 공격패턴 수정

+++ NovaProject/Assets/Scripts/Manager/SpawnManager.cs
@@ -24,7 +24,7 @@ public class SpawnEventData
     public Vector2 direction;
 
     [Tooltip("몬스터 타입 / 패턴 분기용")]
-    public int type;
+    public float type;
 
     [Header("반복 설정")]
     [Tooltip("몇 번 반복할지")]
@@ -146,7 +146,7 @@ public class SpawnManager : Singleton<SpawnManager>
     /// 이 함수를 만든 이유는 오브젝트 풀링 특성상 한번 쓰고난 오브젝트는 위치값이 이상하게 들어가 있어서
     /// 반드시 위치값을 넣어줘야하는데(Instantiate처럼 프리펩 초기값으로 생성되지 않음)
     /// 언젠가 한번은 까먹고 안할것같아서 함수로 묶어뒀습니다
-    private void SpawnMonster(GameObject monster, Vector3 pos, Vector2 dir, int type)
+    private void SpawnMonster(GameObject monster, Vector3 pos, Vector2 dir, float type)
     {
         var go = PoolManager.instance.Get(monster);    // 오브젝트 풀링에서 생성 Instantiate대신 쓴다고 생각하면 될것같습니다
         go.GetComponent<Monster>().Init(pos, dir, type); // 몬스터 안에서 OnEnable로 초기화 할 수 있긴 하지만
